progress
deleted old artBoard files and replaced with artBoard-test
improved styling
