+++ pages/artBoard-page.js
@@ -1,14 +1,14 @@
 import dynamic from 'next/dynamic'
 import React from 'react'
 import axios from 'axios'
-const ArtBoard = dynamic(() => import('../components/artBoard'), {
-  ssr: false,
+const ArtBoard = dynamic(() => import('../components/ArtBoard/index'), {
+    ssr: false
 })
 
 export default function ArtBoardPage() {
-  return (
-    <div>
-      <ArtBoard />
-    </div>
-  )
+    return (
+        <div>
+            <ArtBoard />
+        </div>
+    )
 }
